Monotone-Parent: e1d77c8035a84435f5745fe6e47f1fb5c978d808 Monotone-Revision: 872b6482fb1a591d12d1baf86c6724907aa743dc
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2011-03-21T00:41:22
Monotone-Branch: ca.inverse.sogo

+++ OpenChange/MAPIStoreCalendarMessageTable.m
@@ -44,7 +44,7 @@
   id value;
 
   value = NSObjectFromMAPISPropValue (&res->lpProp);
-  switch (res->ulPropTag)
+  switch ((uint32_t) res->ulPropTag)
     {
     case PR_MESSAGE_CLASS_UNICODE:
       if ([value isEqualToString: @"IPM.Appointment"])
